Extract another class
